Add window created in GUIBuilder
The window contains text, a radio buttons, checkboxes, progbar and buttons.
When an element is clicked, the text at the top of the window displays the name of the clicked element.
Сheckboxes control LEDs on the board.
When you click on the buttons, the value of the progress bar increases or decreases.
An arrow cursor is also displayed.